PUI: Revert to using official pu.h
Becuase we can't rely on the distro PLIB matching what we ship locally, revert to using the
official copy. This may impact people with older distro versions of PLIB, but hopefully in
a way that can be diagnosed.

The original stated reason for including the file ourselves is in #2363 - based on my testing,
adding a self-compiled PLIB to either PLIBDIR or CMAKE_PREFIX_PATH does seem to work
correctly now, using CMake 3.22

SF-ID: https://sourceforge.net/p/flightgear/codetickets/2946/
SF-ID: https://sourceforge.net/p/flightgear/codetickets/2362/