Convert Authors into Markdown.
By suggestion, specific contribution attributed to a name has been removed,
not because it isn't important or valuable, but with the rationale that
it is impossible to even remotely keep the information accurate and up-to-date.
The most important part is that we retain peoples names.

Likewise, email addresses have been removed as:

a) It's impossible to know who still even wants their email made public for spam/privacy purposes
b) There is a very high likelihood that many of these email addresses (and their providers) no longer exist.