tests: Check for runtime errors when running nasal
executeNasal() only checked for parse errors, and left to the caller the
responsibility of checking runtime errors, which almost no tests did.
testAirwaysAPI() and benchPropsNodeSet() were partially broken as a
result: they had runtime errors which were silently ignored.

Change executeNasal() to report failure in case of both parse and
runtime errors, which is what almost all tests need.

Add executeNasalExpectErrors() which distinguishes parse errors and
runtime errors, for the few tests which do expect a runtime error.
So far NasalSysTests::testCommands() is the only such test, convert it
to this new function.